# Allocate uninitialized arrays for faster String operations