--- a/src/java.base/share/classes/java/lang/StringLatin1.java
+++ b/src/java.base/share/classes/java/lang/StringLatin1.java
@@ -27,11 +27,9 @@ package java.lang;
 
 import java.util.Arrays;
 import java.util.Locale;
-import java.util.Objects;
 import java.util.Spliterator;
 import java.util.function.Consumer;
 import java.util.function.IntConsumer;
-import java.util.stream.IntStream;
 import java.util.stream.Stream;
 import java.util.stream.StreamSupport;
 import jdk.internal.util.ArraysSupport;
@@ -747,7 +745,7 @@ final class StringLatin1 {
     }
 
     public static byte[] toBytes(int[] val, int off, int len) {
-        byte[] ret = new byte[len];
+        byte[] ret = StringConcatHelper.newArray(len);
         for (int i = 0; i < len; i++) {
             int cp = val[off++];
             if (!canEncode(cp)) {