@stsypanov 2020-11-03
@shipilev do you think this could be helpful or should I close this one?