@stsypanov 2020-11-19
> I am tentatively against doing this. There are security implications for misusing uninitialized arrays -- it reveals the heap data that is usually protected by Java rules. If there is a code that has a bug or the code that relies on default values for the array, such local issue suddenly becomes the security problem.

JDK already uses `Unsafe.allocateUninitializedArray()` in `StringConcatHelper.newArray()` so I'm just expanding its usage over other relevant places, all of them are within String-related classes.

> Additionally, using more classes from `String` invites bootstrapping circularities. It would be cleaner to use `Unsafe` from `String` to avoid intermediate classes, but even then I am not sure `Unsafe` and `String` are not in bootstrap cycle. `String` is certainly a primordial class.

`StringConcatHelper` is already used in `String`and `StringLatin1` see e.g. `String.concat()`, `StringLatin1.replace()` methods, and `Unsafe` is already inside of `StringConcatHelper`. The only place where we add a new dependecy is `StringUTF16` I believe.